+++ src/screens/bottom/AreaScreen.js
